Trait arrow2_convert::deserialize::TryIntoCollection

pub trait TryIntoCollection<Collection, Element> where
    Element: ArrowField,
    Collection: FromIterator<Element>, {
    fn try_into_collection(self) -> Result<Collection>;
    fn try_into_collection_as_type<ArrowType>(self) -> Result<Collection>
    where
        ArrowType: ArrowDeserialize + ArrowField<Type = Element> + 'static,
        for<'b> &'b <ArrowType as ArrowDeserialize>::ArrayType: IntoIterator;
}

Top-level API to deserialize from Arrow

Required methods

fn try_into_collection(self) -> Result<Collection>

Convert from a arrow2::Array to any collection that implements the FromIterator trait

fn try_into_collection_as_type<ArrowType>(self) -> Result<Collection> where
ArrowType: ArrowDeserialize + ArrowField<Type = Element> + 'static,
for<'b> &'b <ArrowType as ArrowDeserialize>::ArrayType: IntoIterator,

Same as try_into_collection except can coerce the conversion to a specific Arrow type. This is useful when the same rust type maps to one or more Arrow types for example LargeString.