Module arrow2_convert::serialize

Implementation and traits for serializing to Arrow.

Traits

ArrowSerialize

Trait that is implemented by all types that are serializable to Arrow.

FlattenChunk

API to flatten a Chunk consisting of an arrow2::array::StructArray into a Chunk consisting of arrow2::array::Arrays contained by the StructArray

TryIntoArrow

Top-level API to serialize to Arrow

Functions

arrow_serialize_to_mutable_array

Serializes an iterator into an arrow2::MutableArray