Trait arrow2_convert::serialize::FlattenChunk

pub trait FlattenChunk {
    fn flatten(self) -> Result<Chunk<Box<dyn Array>>, Error>;
}

API to flatten a Chunk consisting of an arrow2::array::StructArray into a Chunk consisting of arrow2::array::Arrays contained by the StructArray

Required Methods

fn flatten(self) -> Result<Chunk<Box<dyn Array>>, Error>

Convert an arrow2::chunk::Chunk containing a arrow2::array::StructArray to an arrow2::chunk::Chunk consisting of the arrow::array::Arrays contained by the StructArray by consuming the original Chunk. Returns an error if the Chunk cannot be flattened.