Trait arrow2_convert::serialize::ArrowSerialize

pub trait ArrowSerialize: ArrowField {
    type MutableArrayType: ArrowMutableArray;
    fn new_array() -> Self::MutableArrayType;
    fn arrow_serialize(
        v: &<Self as ArrowField>::Type, 
        array: &mut Self::MutableArrayType
    ) -> Result<()>;
}

Trait that is implemented by all types that are serializable to Arrow.

Implementations are provided for all built-in arrow types as well as Vec, and Option if T implements ArrowSerialize.

Note that Vec implementation needs to be enabled by the crate::arrow_enable_vec_for_type macro.

Associated Types

type MutableArrayType: ArrowMutableArray

The arrow2::array::MutableArray that holds this value

Required methods

fn new_array() -> Self::MutableArrayType

Create a new mutable array

fn arrow_serialize(
v: &<Self as ArrowField>::Type,
array: &mut Self::MutableArrayType
) -> Result<()>

Serialize this field to arrow