Trait arrow2_convert::serialize::TryIntoArrow

pub trait TryIntoArrow<'a, ArrowArray, Element>where
    Self: IntoIterator<Item = &'a Element>,
    Element: 'static,{
    fn try_into_arrow(self) -> Result<ArrowArray>;
    fn try_into_arrow_as_type<ArrowType>(self) -> Result<ArrowArray>
    where
        ArrowType: ArrowSerialize + ArrowField<Type = Element> + 'static;
}

Top-level API to serialize to Arrow

Required Methods

fn try_into_arrow(self) -> Result<ArrowArray>

Convert from any iterable collection into an arrow2::Array

fn try_into_arrow_as_type<ArrowType>(self) -> Result<ArrowArray>where
ArrowType: ArrowSerialize + ArrowField<Type = Element> + 'static,

Convert from any iterable collection into an arrow2::Array by coercing the conversion to a specific Arrow type. This is useful when the same rust type maps to one or more Arrow types for example LargeString.